% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{parallel_accuracy}
\alias{parallel_accuracy}
\title{Accuracy}
\usage{
parallel_accuracy(.mbl, ...)
}
\arguments{
\item{.mbl}{A mable or list of mables}

\item{...}{Parameters to pass to \code{fabletools::accuracy()}}
}
\description{
Get accuracies in parallel
}
