% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{parallel_forecast}
\alias{parallel_forecast}
\title{Forecast}
\usage{
parallel_forecast(.mbl, ...)
}
\arguments{
\item{.mbl}{A mable or list of mables}

\item{...}{Parameters to pass to \code{fabletools::forecast()}}
}
\description{
Forecast in parallel
}
