% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{parallel_model}
\alias{parallel_model}
\title{Model}
\usage{
parallel_model(.ts, ...)
}
\arguments{
\item{.ts}{A tsibble}

\item{...}{Parameters to pass to \code{fabletools::model()}}
}
\description{
Model in parallel
}
